+++ src/components/course/enrollment/common.jsx
@@ -1,4 +1,3 @@
-/* eslint-disable import/prefer-default-export */
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Button } from '@edx/paragon';
