+++ src/components/course/enrollment/constants.js
@@ -7,7 +7,6 @@ const enrollButtonTypesLocal = {
   HIDE_BUTTON: 'hide_button',
 };
 
-// eslint-disable-next-line import/prefer-default-export
 export const enrollButtonTypes = Object.freeze(enrollButtonTypesLocal);
 
 export const enrollLinkClass = 'btn-block';
