+++ src/components/dashboard/main-content/course-enrollments/course-cards/email-settings/data/service.js
@@ -1,7 +1,6 @@
 import { getAuthenticatedHttpClient } from '@edx/frontend-platform/auth';
 import { getConfig } from '@edx/frontend-platform/config';
 
-// eslint-disable-next-line import/prefer-default-export
 export const updateEmailSettings = (courseRunId, hasEmailsEnabled) => {
   const config = getConfig();
   const queryParams = new URLSearchParams({
