+++ src/components/dashboard/main-content/course-enrollments/course-cards/email-settings/index.js
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as EmailSettingsModal } from './EmailSettingsModal';
