+++ src/components/dashboard/main-content/course-enrollments/course-cards/mark-complete-modal/index.js
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as MarkCompleteModal } from './MarkCompleteModal';
