+++ src/components/dashboard/main-content/course-enrollments/course-cards/move-to-in-progress-modal/index.js
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as MoveToInProgressModal } from './MoveToInProgressModal';
