+++ src/components/dashboard/main-content/course-enrollments/data/hooks.js
@@ -9,7 +9,6 @@ import * as service from './service';
 import { groupCourseEnrollmentsByStatus, transformCourseEnrollment } from './utils';
 import { COURSE_STATUSES } from './constants';
 
-// eslint-disable-next-line import/prefer-default-export
 export const useCourseEnrollments = ({
   enterpriseUUID,
   requestedCourseEnrollments,
