+++ src/components/dashboard/main-content/data/constants.js
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export const NUM_POPULAR_COURSES_TO_DISPLAY = 4;
