+++ src/components/enterprise-invite/index.js
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as EnterpriseInvitePage } from './EnterpriseInvitePage';
