+++ src/components/enterprise-user-subsidy/data/constants.js
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/prefer-default-export
 export const LICENSE_STATUS = {
   ACTIVATED: 'activated',
   ASSIGNED: 'assigned',
