+++ src/components/integration-warning-modal/index.js
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export { default as IntegrationWarningModal } from './IntegrationWarningModal';
