+++ src/components/pathway/constants.js
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/prefer-default-export
 export const MAX_VISIBLE_SKILLS_PATHWAY = 4;
 export const PATHWAY_SEARCH_EVENT_NAME = 'edx.ui.enterprise.learner_portal.search.pathway.card.clicked';
 export const PATHWAY_SKILL_QUIZ_EVENT_NAME = 'edx.ui.enterprise.learner_portal.skills_quiz.pathway.card.clicked';
