+++ src/components/program-progress/ProgramProgressCourses.jsx
@@ -1,5 +1,4 @@
 import React, { useContext } from 'react';
-
 import PropTypes from 'prop-types';
 import {
   Form, Col, Row,
@@ -7,19 +6,34 @@ import {
 import moment from 'moment';
 import { CheckCircle } from '@edx/paragon/icons';
 import { AppContext } from '@edx/frontend-platform/react';
+
+import { UserSubsidyContext } from '../enterprise-user-subsidy';
+import { SubsidyRequestsContext } from '../enterprise-subsidy-requests';
 import {
   courseUpgradationAvailable,
   getCertificatePriceString,
   getEnrolledCourseRunDetails,
   getNotStartedCourseDetails,
+  hasLicenseOrCoupon,
 } from './data/utils';
-import { useHasLicenseOrCoupon } from './data/hooks';
 import { NotCurrentlyAvailable } from './data/constants';
 
 const ProgramProgressCourses = ({ courseData }) => {
   const { enterpriseConfig } = useContext(AppContext);
+  const {
+    subscriptionPlan,
+    subscriptionLicense,
+    couponCodes: { couponCodesCount },
+  } = useContext(UserSubsidyContext);
 
-  const hasLicenseOrCoupon = useHasLicenseOrCoupon();
+  const { requestsBySubsidyType } = useContext(SubsidyRequestsContext);
+
+  const userHasLicenseOrCoupon = hasLicenseOrCoupon({
+    requestsBySubsidyType,
+    subscriptionPlan,
+    subscriptionLicense,
+    couponCodesCount,
+  });
 
   let coursesCompleted = [];
   let coursesInProgress = [];
@@ -40,7 +54,7 @@ const ProgramProgressCourses = ({ courseData }) => {
 
   const getCertificatePrice = (course) => {
     const certificatePrice = getCertificatePriceString(course);
-    if (hasLicenseOrCoupon) {
+    if (userHasLicenseOrCoupon) {
       return (
         <>
           {certificatePrice
