+++ src/components/program-progress/data/constants.js
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/prefer-default-export
 export const NO_PROGRAMS_ERROR_MESSAGE = 'You are not enrolled in any programs yet.';
 
 // ProgramProgressCircle constants
