+++ src/components/program-progress/data/utils.jsx
@@ -1,11 +1,15 @@
 import moment from 'moment';
+
+import { SUBSIDY_TYPE } from '../../enterprise-subsidy-requests';
+import { LICENSE_STATUS } from '../../enterprise-user-subsidy/data/constants';
+import { PROGRAM_TYPE_MAP } from '../../program/data/constants';
+
 import MicroMastersProgramDetailsSvgIcon from '../../../assets/icons/micromasters-program-details.svg';
 import ProfCertProgramDetailsSvgIcon from '../../../assets/icons/professional-certificate-program-details.svg';
 import XSeriesProgramDetailsSvgIcon from '../../../assets/icons/xseries-program-details.svg';
 import progCertMicroMaster from '../images/program-certificate-micromasters.gif';
 import progCertProfessionalCert from '../images/program-certificate-professional-certificate.gif';
 import progCertXSeries from '../images/program-certificate-xseries.gif';
-import { PROGRAM_TYPE_MAP } from '../../program/data/constants';
 
 export function getProgramIcon(type) {
   switch (type) {
@@ -182,4 +186,20 @@ export function getCoursesEnrolledInAuditMode(courses) {
     ))
   ));
   return courseRuns;
+}
+
+export function hasLicenseOrCoupon({
+  subscriptionPlan,
+  requestsBySubsidyType,
+  subscriptionLicense,
+  couponCodesCount,
+}) {
+  const licenseRequests = requestsBySubsidyType[SUBSIDY_TYPE.LICENSE];
+  const couponCodeRequests = requestsBySubsidyType[SUBSIDY_TYPE.COUPON];
+
+  const hasActiveLicenseOrLicenseRequest = (subscriptionPlan
+    && subscriptionLicense?.status === LICENSE_STATUS.ACTIVATED) || licenseRequests.length > 0;
+  const hasAssignedCodesOrCodeRequests = couponCodesCount > 0 || couponCodeRequests.length > 0;
+
+  return hasActiveLicenseOrLicenseRequest || hasAssignedCodesOrCodeRequests;
 }
