+++ src/components/program/data/hooks.jsx
@@ -4,7 +4,6 @@ import { camelCaseObject } from '@edx/frontend-platform/utils';
 
 import ProgramService from './service';
 
-// eslint-disable-next-line import/prefer-default-export
 export function useAllProgramData({ enterpriseUuid, programUuid }) {
   const [programData, setProgramData] = useState();
   const [fetchError, setFetchError] = useState();
