+++ src/components/search/popular-results/data/constants.js
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export const NUM_RESULTS_TO_DISPLAY = 4;
