+++ src/components/skills-quiz/data/constants.js
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/prefer-default-export
 export const SET_KEY_VALUE = 'SET_KEY_VALUE';
