+++ src/components/skills-quiz/tests/utils.jsx
@@ -3,7 +3,7 @@ import { renderWithRouter } from '@edx/frontend-enterprise-utils';
 
 import { SearchData } from '@edx/frontend-enterprise-catalog-search';
 import { SkillsContextProvider } from '../SkillsContextProvider';
-// eslint-disable-next-line import/prefer-default-export
+
 export const renderWithSearchContext = (children) => renderWithRouter(
   <SearchData>
     <SkillsContextProvider>
