+++ src/components/utils/hooks.js
@@ -1,6 +1,5 @@
 import { useRef, useEffect } from 'react';
 
-// eslint-disable-next-line import/prefer-default-export
 export const useIsFirstRender = () => {
   const isMountRef = useRef(true);
   useEffect(() => {
